mountcb was added (TL)
mountcb.m adds an existing CellBase to Matlab. It differs from initcb.m
insofar it is not trying to create a new CellBase.mat and it doesn't
search the directory for new cells. initcb.m was not able to handle the
problem when you just want to use CellBase functions with a new computer
on existing data.

mountcb.m was created by Tamas Laszlovszky.